requirements.txt
# external requirements
importlib-metadata<5.0
flake8
python-dotenv>=0.5.1
pytest

numpy
pandas
pyro-ppl
scikit-learn
torch
matplotlib
hydra-core
hydra_colorlog
pytorch-lightning
mlflow
ray[tune]
pyreadr

git+https://github.com/fadel/pytorch_ema